+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+### Release v2.1.6
+
+### Feature
+- Big token movers with subscription [#1191](https://github.com/EthVM/EthVM/pull/1191)
+
+###  Fix
+- fis loading state on portfolio list [#1191](https://github.com/EthVM/EthVM/pull/1191)
+
+### UI
+- clean up color system [#1191](https://github.com/EthVM/EthVM/pull/1191)
+
+======================================================================
+
 ### Release v2.1.5
 
 ### Feature